module: compile many files in one file

+++ diu/conf.h
@@ -1,6 +1,6 @@
 #ifndef CONF_H_
 #define CONF_H_
 
-constexpr int ENGINE_VERSION = 1;
+constexpr double ENGINE_VERSION = 1;
 
 #endif
